+++ base/src/testFixtures/java/eu/fraho/spring/securityJwt/base/it/spring/SecuredController.java
@@ -8,6 +8,9 @@ package eu.fraho.spring.securityJwt.base.it.spring;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,5 +27,17 @@ public class SecuredController {
     @Secured("ROLE_ADMIN")
     public ResponseEntity<String> testAdmin() {
         return ResponseEntity.ok("Hello world!");
+    }
+
+    @RequestMapping("/authentication")
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<String> authentication(Authentication authentication) {
+        if (authentication == null) {
+            return ResponseEntity.ok("null token");
+        } else if (authentication instanceof AnonymousAuthenticationToken) {
+            return ResponseEntity.ok("anonymous token");
+        } else {
+            return ResponseEntity.ok("authenticated token");
+        }
     }
 }